requirements.txt
attrs==23.1.0
autopep8==2.0.4
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
gitdb==4.0.10
GitPython==3.1.37
google-api-core==2.12.0
google-auth==2.23.2
google-cloud-aiplatform==1.34.0
google-cloud-bigquery==3.11.4
google-cloud-bigquery-storage==2.22.0
google-cloud-core==2.3.3
google-cloud-functions==1.13.3
google-cloud-language==2.11.1
google-cloud-storage==2.11.0
google-cloud-translate==3.12.0
google-cloud-vision==3.4.4
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpc-google-iam-v1==0.12.6
grpcio==1.58.0
grpcio-status==1.58.0
idna==3.4
iniconfig==2.0.0
numpy==1.26.0
packaging==23.1
parameterized==0.9.0
pluggy==1.3.0
proto-plus==1.22.3
protobuf==4.24.3
py==1.11.0
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pyparsing==3.1.1
pytest==7.4.2
pytest-json-report==1.5.0
pytest-metadata==3.0.0
python-dateutil==2.8.2
requests==2.31.0
rsa==4.9
six==1.16.0
smmap
toml==0.10.2
tomli==2.0.1
urllib3==2.0.5
